Update radio.c
Line 66 RADIO_RX_INTERRUPTS was expanded to include the IRSSI value which is 1 when an RSSI value it read and this RSSI value is above the threshold defined by _______

In the function radio_configure I made changes to enable the RSSI interrupt. See the line
register_write(EZRADIOPRO_INTERRUPT_ENABLE_2, 0x08); The enables the RSSI interrupt. For details see ENRSSI in interrupt enable 2 in the Datasheet AN440 Detailed Register Descriptions. 
In the radio_configure function I also turned the RSSI THRESHOLD down to 0. This threshold is described in AN440 and can be found by searching the document for RSSITH. This value determines the minimum value a read RSSI must be in order to trigger the interrupt. 
register_write(EZRADIO_RSSI_THRESHOLD,0x00);

finally I added the capability to receive the RSSI values when the interrupt is triggered in the INTERRUPT_INT0 service routine. I have left a comment to designate my code.